+++ src/main/java/com/mawkun/laundry/controller/KindController.java
@@ -3,9 +3,11 @@ package com.mawkun.laundry.controller;
 import cn.pertech.common.abs.BaseController;
 import cn.pertech.common.spring.JsonResult;
 import com.github.pagehelper.PageInfo;
+import com.mawkun.laundry.base.data.UserSession;
 import com.mawkun.laundry.base.data.query.KindQuery;
 import com.mawkun.laundry.base.entity.Kind;
 import com.mawkun.laundry.service.KindServiceExt;
+import com.mawkun.laundry.spring.annotation.LoginedAuth;
 import com.xiaoleilu.hutool.convert.Convert;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -14,6 +16,7 @@ import net.sf.cglib.core.Transformer;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.multipart.MultipartFile;
+import springfox.documentation.annotations.ApiIgnore;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -61,7 +64,8 @@ public class KindController extends BaseController {
 
     @PostMapping("/insert")
     @ApiOperation(value="添加商品类型", notes="添加商品类型")
-    public JsonResult insert(Kind kind, MultipartFile file){
+    public JsonResult insert(@LoginedAuth @ApiIgnore UserSession session, Kind kind, MultipartFile file){
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权添加商品分类");
         if(file == null || kind.getKindName() == null) return sendError("缺少参数");
         Kind resultKind = kindServiceExt.getByName(kind.getKindName());
         if(resultKind != null) return sendError("该类型已存在不能重复添加");
@@ -71,22 +75,25 @@ public class KindController extends BaseController {
 
     @PutMapping("/update")
     @ApiOperation(value="编辑商品类型", notes="编辑商品类型")
-    public JsonResult update(Kind kind, MultipartFile file){
+    public JsonResult update(@LoginedAuth @ApiIgnore UserSession session,Kind kind, MultipartFile file){
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权编辑商品分类");
         int result = kindServiceExt.updateWithPic(kind, file);
         return sendSuccess(result);
     }
 
     @DeleteMapping("/delete")
     @ApiOperation(value="删除商品类型", notes="删除商品类型")
-    public JsonResult deleteOne(Long id){
+    public JsonResult deleteOne(@LoginedAuth @ApiIgnore UserSession session,Long id){
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权删除商品分类");
         int result = kindServiceExt.deleteById(id);
         return sendSuccess(result);
     }
 
     @DeleteMapping("/deleteBatch")
     @ApiOperation(value="批量删除商品类型", notes="批量删除商品类型")
-    public JsonResult deleteBatch(String ids){
+    public JsonResult deleteBatch(@LoginedAuth @ApiIgnore UserSession session,String ids){
         int result = 0;
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权删除商品分类");
         List<String> idArray = Arrays.asList(ids.split(","));
         List idList = new ArrayList<>();
         idList = CollectionUtils.transform(idArray, new Transformer() {
