+++ src/main/java/com/mawkun/laundry/controller/ShopController.java
@@ -6,11 +6,13 @@ import cn.pertech.common.utils.RequestUtils;
 import com.alibaba.fastjson.JSONArray;
 import com.github.pagehelper.PageInfo;
 import com.mawkun.laundry.base.data.ShopIncomeData;
+import com.mawkun.laundry.base.data.UserSession;
 import com.mawkun.laundry.base.data.query.ShopIncomeQuery;
 import com.mawkun.laundry.base.data.query.ShopQuery;
 import com.mawkun.laundry.base.data.query.StateQuery;
 import com.mawkun.laundry.base.entity.Shop;
 import com.mawkun.laundry.service.ShopServiceExt;
+import com.mawkun.laundry.spring.annotation.LoginedAuth;
 import com.xiaoleilu.hutool.convert.Convert;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -21,6 +23,7 @@ import org.apache.commons.lang3.ArrayUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.multipart.MultipartFile;
+import springfox.documentation.annotations.ApiIgnore;
 import sun.security.util.ArrayUtil;
 
 import javax.servlet.http.HttpServletRequest;
@@ -65,14 +68,16 @@ public class ShopController extends BaseController {
 
     @GetMapping("/pageList")
     @ApiOperation(value="门店列表分业", notes="门店列表分业")
-    public JsonResult pageList(ShopQuery shopQuery) {
+    public JsonResult pageList(@LoginedAuth @ApiIgnore UserSession session, ShopQuery shopQuery) {
+        if(session.getShopId() > 0) shopQuery.setId(session.getShopId());
         PageInfo<ShopQuery> page = shopServiceExt.pageByEntity(shopQuery);
         return sendSuccess(page);
     }
 
     @PostMapping("/insert")
     @ApiOperation(value="添加门店", notes="添加门店")
-    public JsonResult insert(Shop shop, MultipartFile[] files){
+    public JsonResult insert(@LoginedAuth @ApiIgnore UserSession session,Shop shop, MultipartFile[] files){
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权添加门店");
         if(files.length == 0) return sendError("请上传门店图片");
         List shopList = shopServiceExt.getByName(shop.getShopName());
         if(!shopList.isEmpty()) return sendError("该门店名称已存在，不能重复添加");
@@ -82,7 +87,8 @@ public class ShopController extends BaseController {
 
     @PutMapping("/update")
     @ApiOperation(value="编辑门店", notes="编辑门店")
-    public JsonResult update(Shop shop, MultipartFile[] files){
+    public JsonResult update(@LoginedAuth @ApiIgnore UserSession session, Shop shop, MultipartFile[] files){
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权编辑门店");
         List shopList = shopServiceExt.getByName(shop.getShopName());
         if(shopList.size() > 1) return sendError("该门店名称已存在不能重复添加");
         int result = shopServiceExt.updateWithPic(shop, files);
@@ -91,15 +97,17 @@ public class ShopController extends BaseController {
 
     @DeleteMapping("/delete")
     @ApiOperation(value="删除门店", notes="删除门店")
-    public JsonResult deleteOne(Long id){
+    public JsonResult deleteOne(@LoginedAuth @ApiIgnore UserSession session,Long id){
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权删除门店");
         int result = shopServiceExt.deleteById(id);
         return sendSuccess(result);
     }
 
     @DeleteMapping("/deleteBatch")
     @ApiOperation(value="批量删除门店", notes="批量删除门店")
-    public JsonResult deleteBatch(String ids){
+    public JsonResult deleteBatch(@LoginedAuth @ApiIgnore UserSession session, String ids){
         int result = 0;
+        if(session.getShopId() > 0) return sendArgsError("子管理员无权删除门店");
         List<String> idArray = Arrays.asList(ids.split(","));
         List idList = new ArrayList<>();
         idList = CollectionUtils.transform(idArray, new Transformer() {
@@ -118,8 +126,9 @@ public class ShopController extends BaseController {
      */
     @GetMapping("/statsShopIncome")
     @ApiOperation(value="统计门店收入", notes="统计门店收入")
-    public JsonResult statsShopIncome() {
+    public JsonResult statsShopIncome(@LoginedAuth UserSession session) {
         StateQuery query = this.createQueryStateVo();
+        if(session.getShopId() > 0) query.setShopId(session.getShopId());
         JSONArray array = shopServiceExt.statsShopIncome(query);
         return sendSuccess(array);
     }
